Add thread pool with sig handlers for cleanup
Add a simple thread pool for accepting requests on the socket. The
threads numbers are implemented using atomic read modify writes on a
shared int. Cleanup handling is done by setting a handler for the
signals SIGTERM and SIGINT which closes the socket and waits for each
thread to die before quitting. The main thread then resumes due to the
`pause` call and exits normally.

The default number of threads is the number of cores, although it's
questionable whether this makes the most sense. The request runtime is
presumably IO bound so it may make sense to increase this, but I saw no
benefit in benchmarks (using ab).

The benchmarks show a marginal increase in performance over the previous
thread-per-connection model. Presumably we could do better by getting
rid of the blocking calls and switching to epoll.